Initial implementation
The main feature of this implementation is `projects sync` command, which allows for the projects from Gitlab instance to be cloned to local machine and updated  later with upstream changes.

Such functionality might be useful when dealing with a bunch of projects, which are updated frequently and one has to get this updates regularly.